[contacts-search] move to child component in Composer, remove contacts-search package
