+++ components/composer/CHANGELOG.md
@@ -6,4 +6,5 @@
 
 ## Bug Fixes
 
-- Don't hide `from` field when account is loaded and add collapsible height to whole composer [#247](https://github.com/nylas/components/pull/247)
+- [Composer] Don't hide `from` field when account is loaded and add collapsible height to whole composer [#247](https://github.com/nylas/components/pull/247)
+- [Composer] Add background color to contact search dropdown items and increase search input width [#251](https://github.com/nylas/components/pull/251)
