--- components/contacts-search/.npmignore
@@ -1,6 +1,0 @@
-# Created by .ignore support plugin (hsz.mobi)
-src
-node_modules
-rollup.config.js
-styles.css
-tsconfig.json
