--- components/contacts-search/bower.json
@@ -1,3 +1,0 @@
-{
-  "name": "@nylas/components-contacts-search"
-}
