--- components/contacts-search/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "@nylas/components-contacts-search",
-  "version": "1.1.3",
-  "repository": {
-    "type": "git",
-    "url": "github:nylas/components.git",
-    "directory": "components/contacts-search"
-  },
-  "scripts": {
-    "build": "rollup -c",
-    "dev": "rollup -c -w",
-    "tscheck": "tsc && svelte-check --compiler-warnings \"missing-custom-element-compile-options:ignore\"",
-    "start": "echo 'Nothing to start'"
-  },
-  "main": "index.js",
-  "gitHead": "cca94159c114ecff613f41caa0a4468e0f698e2c"
-}
