--- components/contacts-search/src/main.ts
@@ -1,4 +1,0 @@
-import "@commons/define-component-patch";
-import ContactsSearch from "./ContactsSearch.svelte";
-
-export default ContactsSearch;
