--- components/contacts-search/src/variables.css
@@ -1,2 +1,0 @@
-.nylas-contacts {
-}
